Fix for issues with extension velocities.
The initial interface values were not using the mean of the current
and neighbor values. In 1D this resulted in

   >>> print skfmm.extension_velocities([-1, -1, 1, 1], [0, 2, 1, 0])
   (array([-1.5, -0.5,  0.5,  1.5]), array([ 1.,  1.,  2.,  2.]))

which is clearly wrong. In addition, the extension velocity
calculation away from the interface was not using the updated distance
function, but the old input valued distance function. It was using
`phi_i` and `phi_` instead of `distance_`. `phi_i` appeared to have
the wrong sign as well, but this wasn't debugged.